Поделил на два проекта
