+++ .idea/compiler.xml
@@ -6,7 +6,8 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="untitled" />
+        <module name="clientexel" />
+        <module name="serverexel" />
       </profile>
     </annotationProcessing>
   </component>
