+++ client-exel/src/main/java/ru/gosha/InputSeeker.java
