+++ client-exel/src/main/java/ru/gosha/SG_Muwa/DetectiveException.java
@@ -1,0 +1,7 @@
+package ru.gosha.SG_Muwa;
+
+public class DetectiveException extends Exception {
+    DetectiveException(String message){
+        super(message);
+    }
+}
