+++ client-exel/src/main/java/ru/gosha/SG_Muwa/SeekerType.java
@@ -1,0 +1,5 @@
+package ru.gosha.SG_Muwa;
+
+public enum SeekerType {
+    Teacher, StudyGroup
+}
