--- src/main/java/ru/gosha/SG_Muwa/DetectiveException.java
@@ -1,7 +1,0 @@
-package ru.gosha.SG_Muwa;
-
-public class DetectiveException extends Exception {
-    DetectiveException(String message){
-        super(message);
-    }
-}
