--- src/main/java/ru/gosha/SG_Muwa/SeekerType.java
@@ -1,5 +1,0 @@
-package ru.gosha.SG_Muwa;
-
-public enum SeekerType {
-    Teacher, StudyGroup
-}
